Moved efiling models to there final habitat; used post_dump decorator to extract summary_lines stuffs
