+++ webservices/common/models/filings.py
@@ -2,8 +2,6 @@ from webservices import docs, utils
 
 from .base import db
 
-from sqlalchemy.ext.hybrid import hybrid_property
-
 
 class Filings(db.Model):
     __tablename__ = 'ofec_filings_mv'
@@ -91,157 +89,3 @@ class EFilings(db.Model):
 
 # TODO: add index on committee id and filed_date
     #  version -- this is the efiling version and I don't think we need this - let's document in API for now, see if there are objections
-
-class BaseFilingSummary(db.Model):
-    __tablename__ = 'real_efile_summary'
-    file_number = db.Column('repid', db.Integer, index=True, primary_key=True)
-    line_number = db.Column('lineno', db.Integer, primary_key=True)
-    column_a = db.Column('cola', db.Float)
-    column_b = db.Column('colb', db.Float)
-
-class BaseFiling(db.Model):
-    __abstract__ = True
-    file_number = db.Column('repid', db.Integer, index=True, primary_key=True)
-    committee_id = db.Column('comid', db.String, index=True, doc=docs.COMMITTEE_ID)
-    coverage_start_date = db.Column('from_date', db.Date)
-    coverage_end_date = db.Column('through_date', db.Date)
-    rpt_pgi = db.Column('rptpgi', db.String, doc=docs.ELECTION_TYPE)
-    report_type = db.Column('rptcode', db.String)
-    image_number = db.Column('imageno', db.Integer)
-    street_1 = db.Column('str1', db.String)
-    street_2 = db.Column('str2', db.String)
-    city = db.Column(db.String)
-    state = db.Column(db.String)
-    zip = db.Column(db.String)
-    election_date = db.Column('el_date', db.Date)
-    election_state = db.Column('el_state', db.String)
-    create_date = db.Column('create_dt', db.Date)
-    sign_date = db.Column(db.Date)
-
-
-class BaseF3PFiling(BaseFiling):
-    __tablename__ = 'real_efile_f3p'
-    file_number = db.Column('repid', db.Integer, index=True, primary_key=True)
-
-    treasurer_last_name = db.Column('lname', db.String)
-    treasurer_middle_name = db.Column('mname', db.String)
-    treasurer_first_name = db.Column('fname', db.String)
-    prefix = db.Column(db.String)
-    suffix = db.Column(db.String)
-    committee_name = db.Column('c_name', db.String, index=True, doc=docs.CANDIDATE_NAME)
-    street_1 = db.Column('c_str1', db.String)
-    street_2 = db.Column('c_str2', db.String)
-    city = db.Column('c_city', db.String)
-    state = db.Column('c_state', db.String)
-    zip = db.Column('c_zip', db.String)
-    total_receipts = db.Column('tot_rec', db.Float)
-    total_disbursements = db.Column('tot_dis', db.Float)
-    cash_on_hand_beginning_period = db.Column('cash', db.Float)
-    cash_on_hand_end_period = db.Column('cash_close', db.Float)
-    debts_owed_to_committee = db.Column('debts_to', db.Float)
-    operating_expenditures = db.Column('expe', db.Float)
-    net_contributions = db.Column('net_con', db.Float)
-    net_operating_expenditures = db.Column('net_op', db.Float)
-    primary_election = db.Column('act_pri', db.String)
-    general_election = db.Column('act_gen', db.String)
-    sub_total_sum = db.Column('sub', db.String)
-
-    @hybrid_property
-    def treasurer_name(self):
-        name = name_generator(self.treasurer_first_name,
-                              ' ',
-                              self.treasurer_middle_name,
-                              ' ',
-                              self.treasurer_last_name
-                              )
-
-        name = (
-            name
-            if name
-            else None
-        )
-        return name
-
-    summary_lines = db.relationship(
-        'BaseFilingSummary',
-        primaryjoin='''and_(
-                BaseF3PFiling.file_number == BaseFilingSummary.file_number,
-            )''',
-        foreign_keys=file_number,
-        uselist=True,
-    )
-
-def name_generator(*args):
-    name = ''
-    for field in args:
-        temp_name = (
-            field
-            if field
-            else ''
-        )
-        name += temp_name
-
-    ret_name = ''
-    for string in name.split(' '):
-        ret_name += string.lower().capitalize() + ' '
-    return ret_name.strip()
-
-class BaseF3Filing(BaseFiling):
-    __tablename__ = 'real_efile_f3'
-    file_number = db.Column('repid', db.Integer, index=True, primary_key=True)
-    candidate_last_name = db.Column('can_lname', db.String)
-    candidate_first_name = db.Column('can_fname', db.String)
-    candidate_middle_name = db.Column('can_mname', db.String)
-    candidate_prefix = db.Column('can_prefix', db.String)
-    candidate_suffix = db.Column('can_suffix', db.String)
-    f3z1 = db.Column(db.Integer)
-    primary_election = db.Column('act_pri', db.String)
-    general_election = db.Column('act_gen', db.String)
-    special_election = db.Column('act_spe', db.String)
-    runoff_election = db.Column('act_run', db.String)
-    district = db.Column('eld', db.Integer)
-    amended_address = db.Column('amend_addr', db.String)
-
-    @hybrid_property
-    def candidate_name(self):
-        name = name_generator(
-                              self.candidate_first_name,
-                              '',
-                              self.candidate_middle_name,
-                              '',
-                              self.candidate_last_name
-                              )
-        name = (
-            name
-            if name
-            else None
-        )
-        return name
-
-    summary_lines = db.relationship(
-        'BaseFilingSummary',
-        primaryjoin='''and_(
-                BaseF3Filing.file_number == BaseFilingSummary.file_number,
-            )''',
-        foreign_keys=file_number,
-        uselist=True,
-    )
-
-class BaseF3XFiling(BaseFiling):
-    __tablename__ = 'real_efile_f3x'
-    file_number = db.Column('repid', db.Integer, index=True, primary_key=True)
-
-    committee_name = db.Column('com_name', db.String, index=True, doc=docs.COMMITTEE_NAME)
-    sign_date = db.Column('date_signed', db.Date)
-    amend_address = db.Column('amend_addr', db.String)
-    qual = db.Column(db.String)
-
-
-    summary_lines = db.relationship(
-        'BaseFilingSummary',
-        primaryjoin='''and_(
-                BaseF3XFiling.file_number == BaseFilingSummary.file_number,
-            )''',
-        foreign_keys=file_number,
-        uselist=True,
-    )
